pyproject.toml
[tool.poetry]
name = "ultimate-pollbot"
version = "1.6.0"
description = "A telegram bot for creating all kinds of polls."
authors = ["Arne Beer <contact@arne.beer>"]
license = "MIT"
repository = "https://github.com/nukesor/ultimate-poll-bot"

[tool.poetry.dependencies]
python = ">=3.9"
sentry-sdk = "^0.19"
alembic = "^1"
python-telegram-bot = "^12"
SQLAlchemy = "^1"
sqlalchemy-utils = "^0.37"
psycopg2-binary = "^2"
python-dateutil = "^2"
toml = "^0.10"
python-i18n = {version = "^0.3", extras = ["yaml"]}

matplotlib = "^3"
pandas = "^1"
pillow = "^8"
argparse = "^1"
"ruamel.yaml" = "^0.16"

[tool.poetry.dev-dependencies]
pytest = "^6"
black = { version = "*", allow-prereleases = true }
typer = "^0.3"
types-toml = "0.1.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
